+++ ft-plugin/src/main/java/com/ft/plugin/garble/ClassNameAnalytics.java
@@ -66,7 +66,9 @@ public class ClassNameAnalytics {
     public static boolean isAndroidPackage(String className) {
         return (className.startsWith("android/")
                 || className.startsWith("androidx/"))
-                || className.startsWith("com/google/android");
+                || className.startsWith("com/google/android")
+                || className.startsWith("java/")
+                || className.startsWith("javax/");
     }
 
     public static boolean isAndroidGenerated(String className){